+++ packages/web-visualization/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.17 (2/4/2026 PST)
+
+#### 🚀 Updates
+
+- Add support preferred side for scrubber beacon label group. [[#366](https://github.com/coinbase/cds/pull/366)]
+
 ## 3.4.0-beta.16 (1/29/2026 PST)
 
 #### 🚀 Updates